requirements.txt
PyQt6>=6.6.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
numpy>=1.26.0
Pillow>=10.1.0
pygltflib>=1.16.0
trimesh>=4.0.5
